fix(parser): performance improvements
@axefrog You mentioned before on the Cherow repo that it maybe should be easy to extend the parser?

Well. I have to admit now after one day in h*** that changing any bit masks in this parser
provides unforeseen side effects resulting in at least 2000 broken edge case tests. And it took me 1 day to track and fix it :( :(  And when I finaly fixed it, well 450  new broken tests :(

Anyway. The base layer is laid for #21 , but the language itself have limitations so I'm not sure how to fix that either.