Support Node 6
The "engines" entry in package.json was introduced in commit
38096e50, as part of an eslint configuration cleanup, but it could
have been refined to only apply to the scripts directory. as the rest
of the code can run and pass tests on Node.js 6.

This patch refines the "node/no-unsupported-features/es-syntax" rule
to apply Node.js 10 version to the scripts dir, and relaxes the
requirement of the rest of the project to Node.js 6, so it can be used
in older projects.

CI jobs that run the tests on Node 6 and Node 8 were also added to
make sure these assumptions are correct, and won't be broken in the
future (at least, not without an explicit decision to drop this
support by removing these CI jobs).